requirements.txt
aiohttp==3.6.3
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.2.0
Brlapi==0.6.6
cachetools==4.1.1
catfish==1.4.4
certifi==2020.6.20
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
discord.py==1.5.1
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.1
menulibre==2.2.0
mugshot==0.4.0
multidict==4.7.6
netifaces==0.10.4
numpy==1.17.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
powerline-status==2.6
protobuf==3.13.0
protonvpn-cli==2.2.2
psutil==5.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.0
python-debian==0.1.32
python-debianbts==2.7.2
python-dotenv==0.15.0
python-magic==0.4.16
pythondialog==3.5.1
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.6
SecretStorage==2.3.1
selenium==3.141.0
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.15.0
soupsieve==2.0.1
system-service==0.3
systemd-python==234
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
xcffib==0.5.1
xkit==0.0.0
yarl==1.5.1
zope.interface==4.3.2
